feat: starting session->acquisition merge
